Update Warehouse Management API untuk mendukung arsitektur API Gateway. Perubahan fokus pada transisi autentikasi ke JWT (Zero Trust) dan sinkronisasi identitas user ke Audit Service via gRPC.
1. **Security & Authentication Migration:**
   - Mengganti sistem **X-API-KEY** dengan **JWT (Bearer Token)**.
   - Update `AuthMiddleware` untuk mengekstrak identitas user (ID, Username, Role) dari claims token.
   - Injeksi identitas ke dalam context untuk digunakan di layer service dan logging.

2. **Audit Service Traceability:**
   - Refaktor signature `UpdateStock` pada `ItemService` agar menerima parameter `token`.
   - Implementasi **Token Propagation**: Token dari Gateway diteruskan ke Warehouse, lalu diteruskan lagi ke Audit Service via gRPC metadata.

3. **Documentation (Swagger):**
   - Update `securityDefinitions` menjadi `BearerAuth`.
   - Update semua tag `@Security` pada handler Item dan Category.
   - Penambahan instruksi eksplisit di deskripsi Swagger tentang cara mendapatkan token dari Auth Service.

4. **Code Quality & Fixes:**
   - Perbaikan bug pada `main.go` terkait jumlah parameter pemanggilan `RoleMiddleware`.
   - Update `MockItemService` dan sinkronisasi seluruh Unit Test (`item_handler_test.go` & `item_service_test.go`) agar tetap hijau (pass).

+++ internal/repository/item_mock.go
@@ -35,7 +35,7 @@ func (m *MockItemRepo) Update(id int, item entity.UpdateItemRequest) error {
 	return m.Called(id, item).Error(0)
 }
 
-func (m *MockItemRepo) UpdateStock(req entity.UpdateStockRequest, userID int) error {
+func (m *MockItemRepo) UpdateStock(req entity.UpdateStockRequest, userID int64) error {
 	return m.Called(req, userID).Error(0)
 }
